login_register/requirements.txt
amqp==2.5.1
astroid==2.3.1
billiard==3.6.1.0
celery==4.3.0
Django==2.2.6
django-ranged-response==0.2.0
django-simple-captcha==0.5.12
entrypoints==0.3
flake8==3.7.8
image==1.5.27
importlib-metadata==0.23
isort==4.3.21
jedi==0.15.1
kombu==4.6.5
lazy-object-proxy==1.4.2
mccabe==0.6.1
more-itertools==7.2.0
parso==0.5.1
Pillow==6.2.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.2
pytz==2019.3
qrcode==6.1
redis==3.3.10
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
vine==1.3.0
wrapt==1.11.2
zipp==0.6.0
